# docs/swagger.yaml
definitions:
  image.FileItem:
    properties:
      name:
        description: 文件或目录名
        type: string
      path:
        description: 相对路径
        type: string
      type:
        description: file | dir
        type: string
      url:
        description: 文件访问 URL，仅 file 有
        type: string
    type: object
  image.ListFilesResp:
    properties:
      code:
        $ref: '#/definitions/response.Code'
      data:
        items:
          $ref: '#/definitions/image.FileItem'
        type: array
      msg:
        type: string
    type: object
...
  response.Code:
    enum:
    - 0
    - 200
    - 401
    - 500
    - 501
    - 502
    - 503
    type: integer
    x-enum-varnames:
    - Code_Common
    - Code_Success
    - Code_Unauthorized
    - Code_Err
    - Code_DBErr
    - Code_PasswordErr
    - Code_AlreadyExists
  response.CommonResp:
    properties:
      code:
        $ref: '#/definitions/response.Code'
      msg:
        type: string
    type: object
info:
  contact:
    name: buyfakett
    url: https://github.com/buyfakett
paths:
  /api/image/delete:
    delete:
      consumes:
      - application/json
      description: 删除相对图片目录下的文件
      parameters:
      - description: 文件路径
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResp'
      summary: 删除文件
      tags:
      - 图片
  /api/image/list:
    get:
      consumes:
      - application/json
      description: 获取指定目录下的文件与子目录（不递归）
      parameters:
      - description: 相对 图片 的目录，不传表示根目录
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/image.ListFilesResp'
      summary: 获取文件列表
      tags:
      - 图片
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"